requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
sentence-transformers>=2.3.0
faiss-cpu>=1.13.0
rank-bm25>=0.2.2
transformers>=4.37.0
torch>=2.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.26.0,<2.0.0
tqdm>=4.66.0
openai>=1.10.0
python-dotenv>=1.0.0
requests>=2.31.0

# Visualization dependencies
plotly>=5.18.0
pandas>=2.1.0
scikit-learn>=1.3.0
umap-learn>=0.5.5
kaleido>=0.2.1

# Tokenier integration dependencies
xgboost>=2.0.0
joblib>=1.3.0
regex>=2023.0.0
